Skip non I-frames at startup and after seek for RTSP
When seeking or starting an RTSP stream the video might be blurry
until the first I-frame is rendered.
This fix makes sure that the first rendered frame at startup and
after seek is an I-frame (for AVC video).
If an access unit containing SPS information arrives, the decoding
is also started.

Change-Id: I236bd97542f313a3ddcb27c8adf86d25f3f570d8