Introduce AUDIO_LEGACY_HACK to extend support for legacy devices
this is a dirty hack but does what it should.

if we build for legacy device, eg LG Optimus Black P970, we have to use
the ICS_AUDIO_BLOBS flag to get audio.

but if we add it, we get build breakages and if we remove it, we dont have
audio at all.

this hack allows us to set the AUDIO_LEGACY_HACK flag in the BoardConfig and
prevents build breakages.

as i am unsure if every device needs that, i couldnt wrap it around the
ICS_AUDIO_BLOBS flag, thus this new flag is added.

Change-Id: Id545a684fcdc0508f4135eb7a722a9b617595532