av: stagefright: Fix secure buffers for Widevine DRM playback
CAF added a method to MediaSource which changes the order of the
virtual method "setBuffers" and the destructor.

Objects built prior to this change have a vtable that contains the
destructor in the slot that the framework thinks contains setBuffers.

On secure media playback (e.g. saving a HD video in Google Play
Movies & TV and then playing the downloaded) copy, an attempt it
made to use secure buffers which causes the framework to attempt
to call setBuffers.  The error case will appear in the log as:

E/OMXCodec(  212): [OMX.qcom.video.decoder.avc.secure] Codec requested to use secure input buffers but upstream source didn't support that.

On system that use a Widevine DRM plugin built prior to this
change (all existing commercial Android builds?), this crashes
mediaserver because the attempt to setBuffers actually destroys
the object.

Hack around the change by moving the new method to the end of the
object.

Fixes JIRA-4944

Change-Id: I4ce24a16f3eceab0c4def4b09850ed569eb392ab