Avoid scanning sources when decoder state is shutdown
If creating MediaCodec is failed, decoder notifies error and shutdown
of decoder is requested. In that case, NuPlayer::flushDecoder()
requests performScanSources() if mScanSourcesPending is true. By this
request, kWhatScanSources is posted again and instantiating decoder
is also requested again. This processing is executed continuously and
error is notified many times.
This fix limits running performScanSources() only when needShutdown is
false

Change-Id: I7b7a82ad8fdf493b58ce95670e37e77533295ddc